% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_applyButtonInput.R
\name{mod_applyButtonInput}
\alias{mod_applyButtonInput}
\alias{mod_applyButton}
\title{mod_applyButtonInput and mod_applyButton}
\usage{
mod_applyButtonInput(id)

mod_applyButton(
  input,
  output,
  session,
  lang,
  texts_thes,
  var_thes,
  numerical_thes,
  data_reactives,
  filter_reactives
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{language selected}

\item{texts_thes, var_thes, numerical_thes}{thesauruses}

\item{data_reactives, filter_reactives}{reactives needed}
}
\description{
A shiny module to create and populate the buttons inputs
}
