% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_data}
\alias{mod_data}
\title{mod_data server function}
\usage{
mod_data(input, output, session, nfidb, lang, var_thes, numerical_thes,
  texts_thes)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{nfidb}{lfcdata::lfcNFI object to access the nfi db}

\item{lang}{lang reactive}

\item{var_thes, }{numerical_thes, texts_thes thesauruses}
}
\description{
mod_data server function
}
