% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_dataInput}
\alias{mod_dataInput}
\title{mod_dataInput and mod_data}
\usage{
mod_dataInput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the data inputs
}
