% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataTableOutput.R
\name{mod_dataTableOutput}
\alias{mod_dataTableOutput}
\alias{mod_dataTable}
\title{mod_dataTableOutput and mod_dataTable}
\usage{
mod_dataTableOutput(id)

mod_dataTable(
  input,
  output,
  session,
  main_data_reactives,
  data_reactives,
  viz_reactives,
  nfidb,
  var_thes,
  texts_thes,
  numerical_thes,
  lang
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives}{reactives from
other modules}

\item{nfidb}{pool object to access nfi db}

\item{var_thes, texts_thes, numerical_thes}{thesauruses}

\item{lang}{lang value}
}
\description{
A shiny module to generate and populate the visualization inputs
}
