% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtersUI.R
\name{mod_filtersUI}
\alias{mod_filtersUI}
\alias{mod_filters}
\title{mod_filtersUI and mod_filters}
\usage{
mod_filtersUI(id)

mod_filters(
  input,
  output,
  session,
  lang,
  data_reactives,
  cache,
  var_thes,
  numerical_thes,
  texts_thes,
  categorical_thes
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang reactive value}

\item{data_reactives}{reactives from the dataInput module}

\item{cache}{memoryCache to store the filters values}

\item{var_thes, numerical_thes, texts_thes, categorical_thes}{thesauruses}
}
\description{
A shiny module to generate and process the filters
}
\details{
The logic is the following: With the available variables on the data,
  meaning main table, plots table and dynamic table if any, lets build
  inputs in three categories, results, climatic and other.
  When any variable is selected, create an on the fly input (in a loop), and
  collect the value for the input. This should be cached, this way if the
  user selects a variable and set a value to filter, but after they select
  another variable, the first one set values are remembered.
  After collecting the inputs, we need to build non evaluated filter
  expressions to return, as they will be needed later in the data
  retrieving module.
}
