% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_helpUI.R
\name{mod_helpUI}
\alias{mod_helpUI}
\alias{mod_help}
\title{mod_helpUI and mod_help}
\usage{
mod_helpUI(id)

mod_help(
  input,
  output,
  session,
  data_reactives,
  nfidb,
  var_thes,
  texts_thes,
  numerical_thes,
  lang
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives}{reactives needed}

\item{nfidb}{object to access the database, with describe_var method}

\item{var_thes, texts_thes, numerical_thes}{thesauruses}

\item{lang}{language selected}
}
\description{
module for creating the astounding viz when click
}
