% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoUI.R
\name{mod_info}
\alias{mod_info}
\title{mod_info server function}
\usage{
mod_info(
  input,
  output,
  session,
  map_reactives,
  main_data_reactives,
  viz_reactives,
  nfidb,
  var_thes,
  texts_thes,
  numerical_thes,
  lang
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{map_reactives, main_data_reactives, viz_reactives}{reactives}

\item{nfidb}{object with the db connection}

\item{var_thes, texts_thes, numerical_thes}{thesauruses}

\item{lang}{lang reactive}
}
\description{
mod_info server function
}
