% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainDataOutput.R
\name{mod_mainData}
\alias{mod_mainData}
\title{mod_mainData server function}
\usage{
mod_mainData(input, output, session, data_reactives, filter_reactives,
  apply_reactives, nfidb, lang, texts_thes)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives}{reactives from dataInput module}

\item{filter_reactives}{filter reactives}

\item{nfidb}{object to access the nfi db#'}

\item{lang}{lang selected}

\item{texts_thes}{thesaurus}
}
\description{
mod_mainData server function
}
