% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainDataOutput.R
\name{mod_mainDataOutput}
\alias{mod_mainDataOutput}
\title{mod_mainDataOutput and mod_mainData}
\usage{
mod_mainDataOutput(id)
}
\arguments{
\item{id}{}
}
\description{
Shiny module to get the data as tbl_sql
}
