% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapOutput.R
\name{mod_mapOutput}
\alias{mod_mapOutput}
\alias{mod_map}
\title{mod_mapOutput and mod_map}
\usage{
mod_mapOutput(id)

mod_map(
  input,
  output,
  session,
  data_reactives,
  viz_reactives,
  main_data_reactives,
  lang,
  var_thes,
  texts_thes,
  numerical_thes
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives, viz_reactives, main_data_reactives}{reactives}

\item{lang}{lang selected}

\item{var_thes, texts_thes, numerical_thes}{thesauruses}
}
\description{
Shiny module to generate the map
}
