% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_saveUI.R
\name{mod_saveUI}
\alias{mod_saveUI}
\alias{mod_save}
\title{mod_saveUI and mod_save}
\usage{
mod_saveUI(id)

mod_save(
  input,
  output,
  session,
  map_reactives,
  table_reactives,
  var_thes,
  texts_thes,
  numerical_thes,
  lang
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{map_reactives, table_reactives}{reactives needed}

\item{var_thes, texts_thes, numerical_thes}{thesauruses}

\item{lang}{language selected}
}
\description{
module for creating the astounding viz when click
}
