% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtersUI.R
\name{mod_filtersUI}
\alias{mod_filtersUI}
\alias{mod_filters}
\title{mod_filtersUI and mod_filters}
\usage{
mod_filtersUI(id)

mod_filters(input, output, session, nfidb, lang, data_reactives, var_thes,
  numerical_thes, texts_thes)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{nfidb}{pool object to access the nfi db}

\item{lang}{lang reactive value}

\item{data_reactives}{reactives from the dataInput module}

\item{var_thes, numerical_thes, texts_thes}{thesauruses}

\item{nfidb}{object to access the nfi db}
}
\description{
A shiny module to generate and process the filters
}
