% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainDataOutput.R
\name{mod_mainData}
\alias{mod_mainData}
\title{mod_mainData server function}
\usage{
mod_mainData(
  input,
  output,
  session,
  data_reactives,
  filter_reactives,
  apply_reactives,
  nfidb,
  lang,
  texts_thes,
  parent_session
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives, filter_reactives, apply_reactives}{reactives from modules}

\item{nfidb}{object to access the nfi db}

\item{lang}{lang selected}

\item{texts_thes}{thesaurus}

\item{parent_session}{parent session to be able to update tabset panel}
}
\description{
mod_mainData server function
}
