% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vizInput.R
\name{mod_viz}
\alias{mod_viz}
\title{mod_viz server function}
\usage{
mod_viz(
  input,
  output,
  session,
  data_reactives,
  filter_reactives,
  nfidb,
  var_thes,
  texts_thes,
  numerical_thes,
  categorical_thes,
  lang
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_reactives, filter_reactives}{reactives from
other modules}

\item{nfidb}{pool object to access nfi db}

\item{var_thes, texts_thes, numerical_thes}{thesauruses}

\item{lang}{lang value}
}
\description{
mod_viz server function
}
